factored out utils functions into other areas, some reorganisation and small typos
