+++ asf_heat_pump_suitability/analysis/hn_zones/hnz_utils/hnz_comparison_analysis_utils.py
@@ -10,6 +10,10 @@ DESNZ heat network zones with Nesta's heat network suitability data.
 - Calculates Mean Absolute Error (MAE) between DESNZ pilot scores and Nesta suitability scores.
 
 **Functions:**
+- `setup_paths`: Determines local or S3 file paths based on input flags.
+- `save_gdf_to_gpkg`: Saves a GeoDataFrame as a GeoPackage and uploads it to S3 if needed.
+- `_compute_lsoa_coverage_stats`: Computes spatial intersections and LSOA coverage stats.
+- `load_transform_hn_geodata`: Load the Heat Networks GeoPackage file, ensure consistent CRS, performs a spatial join with LSOA polygons, and calculates the intersection area.
 - `add_DESNZ_pilot_fraction`: Merges heat network zone data with suitability scores and computes the fraction covered.
 - `calculate_average_scores_for_thresholds`: Computes average suitability scores for different thresholds of coverage.
 - `calculate_mae_for_pilot_score`: Calculates MAE for LSOAs inside and outside pilot heat network zones.
@@ -22,6 +26,139 @@ from typing import Tuple, List, Optional
 import geopandas as gpd
 import polars as pl
 import logging
+from asf_heat_pump_suitability.utils.save_utils import upload_file_to_s3
+from asf_heat_pump_suitability.getters.get_datasets import load_desnz_geodata
+from asf_heat_pump_suitability.utils.geo_utils import ensure_crs_match
+from config.hnz_config import (
+    LSOA_SHP_PATH_LOCAL,
+    LSOA_SHP_PATH_S3,
+    NESTA_HPS_PARQUET_LOCAL,
+    NESTA_HPS_PARQUET_S3,
+)
+import geopandas as gpd
+import os
+
+
+def setup_paths(read_in_s3: bool) -> dict:
+    """
+    Set up the paths based on whether to read from S3 or locally.
+
+    Args:
+        read_in_s3 (bool): If True, set up paths to read from S3. Otherwise, set up local paths.
+
+    Returns:
+        dict: A dictionary containing paths for LIVERPOOL_GPKG_PATH, LSOA_SHP_PATH, and NESTA_HP_SUITABILITY_PARQUET_PATH.
+    """
+    paths = {}
+    if read_in_s3:
+        paths["LSOA_SHP_PATH"] = LSOA_SHP_PATH_S3
+        paths["NESTA_HP_SUITABILITY_PARQUET_PATH"] = NESTA_HPS_PARQUET_S3
+
+    else:
+        paths["LSOA_SHP_PATH"] = LSOA_SHP_PATH_LOCAL
+        paths["NESTA_HP_SUITABILITY_PARQUET_PATH"] = NESTA_HPS_PARQUET_LOCAL
+    return paths
+
+
+def save_gdf_to_gpkg(
+    gdf: gpd.GeoDataFrame,
+    output_dir: str,
+    filename_prefix: str,
+    save_to_s3: bool,
+    s3_bucket: str,
+    s3_key_dir: str,
+    subfolder: str,
+):
+    """
+    Save a GeoDataFrame to a GPKG file, then optionally upload to S3.
+
+    Args:
+        gdf (gpd.GeoDataFrame): The GeoDataFrame to write.
+        output_dir (str): Local directory to save the GPKG.
+        filename_prefix (str): The prefix used for naming the GPKG file.
+        save_to_s3 (bool): Whether to upload the output file to S3.
+        s3_bucket (str): Name of the S3 bucket.
+        s3_key_dir (str): The S3 key prefix where files are stored.
+        subfolder (str): Optional subfolder within S3 for organization.
+    """
+    gpkg_filename = (
+        f"{filename_prefix.lower().replace(' ', '_')}_with_desnz_hn_lsoa.gpkg"
+    )
+    gpkg_local_file_path = os.path.join(output_dir, gpkg_filename)
+
+    gdf.to_file(filename=gpkg_local_file_path, driver="GPKG")
+    logging.info(f"Saved GPKG for {filename_prefix} to {gpkg_local_file_path}")
+
+    # s3_key = f"{s3_key_dir}{subfolder}/{gpkg_filename}"
+    upload_file_to_s3(
+        local_file_path=gpkg_local_file_path,
+        s3_bucket=s3_bucket,
+        s3_key_dir=s3_key_dir,
+        filename=gpkg_filename,
+        save_to_s3=save_to_s3,
+        subfolder=subfolder,
+    )
+
+
+def _compute_lsoa_coverage_stats(
+    desnz_hn_gdf: gpd.GeoDataFrame, lsoa_gdf: gpd.GeoDataFrame
+) -> Tuple[gpd.GeoDataFrame, List[str]]:
+    """
+    Perform a spatial intersection between DESNZ heat network zones and LSOA polygons,
+    then calculate the fraction of each LSOA's area that is covered by heat network zones.
+
+    Args:
+        desnz_hn_gdf (gpd.GeoDataFrame): DESNZ heat network zones.
+        lsoa_gdf (gpd.GeoDataFrame): LSOA polygons.
+
+    Returns:
+        Tuple[gpd.GeoDataFrame, List[str]]:
+            - GeoDataFrame containing LSOA codes, intersection geometry, and fraction_covered.
+            - List of unique LSOA codes present in the final GeoDataFrame.
+    """
+    # Remove 'index_right' column if it exists
+    if "index_right" in desnz_hn_gdf.columns:
+        desnz_hn_gdf = desnz_hn_gdf.drop(columns=["index_right"])
+
+    # Ensure LSOAs have a valid total area column
+    lsoa_gdf["total_area"] = lsoa_gdf.geometry.area
+
+    # Perform intersection to determine which LSOAs overlap with heat network zones
+    intersection_gdf = gpd.overlay(desnz_hn_gdf, lsoa_gdf, how="intersection")
+
+    # Calculate the fraction of the LSOA that is covered by heat network zones
+    intersection_gdf["fraction_covered"] = (
+        intersection_gdf["geometry"].area / intersection_gdf["total_area"]
+    )
+
+    # Extract unique LSOA codes
+    unique_lsoa_codes = intersection_gdf["LSOA21CD"].dropna().unique().tolist()
+
+    return intersection_gdf, unique_lsoa_codes
+
+
+def load_transform_hn_geodata(
+    desnz_hn_gpkg_path: str, lsoa_shp_path: str, layer_name: str
+) -> Tuple[gpd.GeoDataFrame, List[str]]:
+    """
+    Load the Heat Networks GeoPackage file, ensure consistent CRS,
+    perform a spatial join with LSOA polygons, and calculate the intersection area.
+
+    Args:
+        desnz_hn_gpkg_path (str): Path to the GeoPackage file with the Heat Network zones.
+        lsoa_shp_path (str): Path to the LSOA shapefile.
+        layer_name (str): Layer name in the GeoPackage.
+
+    Returns:
+        Tuple[gpd.GeoDataFrame, List[str]]:
+            - GeoDataFrame with LSOA codes added via spatial join.
+            - List of unique LSOA codes present in the GeoDataFrame.
+    """
+    desnz_hn_gdf, lsoa_gdf = load_desnz_geodata(
+        gpkg_path=desnz_hn_gpkg_path, shp_path=lsoa_shp_path, layer_name=layer_name
+    )
+    lsoa_gdf = ensure_crs_match(gdf_1=desnz_hn_gdf, gdf_2=lsoa_gdf)
+    return _compute_lsoa_coverage_stats(desnz_hn_gdf, lsoa_gdf)
 
 
 def add_DESNZ_pilot_fraction(
